+++ Cursus/Libft/ft_strrchr.c
@@ -6,14 +6,14 @@
 /*   By: akabbadj <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 00:35:34 by akabbadj          #+#    #+#             */
-/*   Updated: 2022/10/30 12:54:00 by akabbadj         ###   ########.fr       */
+/*   Updated: 2022/11/02 22:51:17 by akabbadj         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int	slen;
+	size_t	slen;
 
 	slen = ft_strlen(s);
 	while (slen >= 0)
